backend/requirements.txt
fastapi
uvicorn[standard]
groq
pytesseract
Pillow
python-multipart
